requirements.txt
amqp==5.0.6
asgiref==3.3.4
autopep8==1.5.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
psycopg2
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
flake8==3.9.1
idna==2.10
importlib-metadata==4.0.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.1.2
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
